TEST Acceptance test to verify we preserve cover info
We should have the same report before loading a moka and after stopping it,
regardless of the calls done while in moked mode.

Note we are ignoring all calls done in moked mode. This is not the ideal
behaviour but is better than losing all the history so we will implement this
first and maybe look into counting calls in moked mode later